DAOS-623 build: Move more scons code into site_tools.
Move build_info in prereqs module.
Move stack analyser into a proper tool
Move AppendIfSupported to compiler code where it's used.
Update GitHub actions triggers.

This all tidies up the build code and means less code needs to
be loaded to make the docker base images leading to more
layer re-use and easier to follow code.

Remove the no-longer used mocked_tests module.
Enable strict pylint checking for build code as it's now
error free.

Required-githooks: true
