+++ site_scons/prereq_tools/base.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2022 Intel Corporation
+# Copyright 2016-2023 Intel Corporation
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,6 +22,9 @@
 
 # pylint: disable=too-many-lines
 import os
+import sys
+import json
+import datetime
 import traceback
 import hashlib
 import time
@@ -30,7 +33,6 @@ import shutil
 import subprocess  # nosec
 import tarfile
 import configparser
-from build_info import BuildInfo
 from SCons.Variables import PathVariable
 from SCons.Variables import EnumVariable
 from SCons.Variables import ListVariable
@@ -492,79 +494,55 @@ class WebRetriever():
             raise UnsupportedCompression(subdir)
 
 
-def check_flag_helper(context, compiler, ext, flag):
-    """Helper function to allow checking for compiler flags"""
-    if compiler in ["icc", "icpc"]:
-        flags = ["-diag-error=10006", "-diag-error=10148", "-Werror-all", flag]
-        # bug in older scons, need CFLAGS to exist, -O2 is default.
-        context.env.Replace(CFLAGS=['-O2'])
-    elif compiler in ["gcc", "g++"]:
-        # pylint: disable=wrong-spelling-in-comment
-        # remove -no- for test
-        # There is a issue here when mpicc is a wrapper around gcc, in that we can pass -Wno-
-        # options to the compiler even if it doesn't understand them but.  This would be tricky
-        # to fix gcc only complains about unknown -Wno- warning options if the compile Fails
-        # for other reasons anyway.
-        test_flag = flag.replace("-Wno-", "-W")
-        flags = ["-Werror", test_flag]
-    else:
-        flags = ["-Werror", flag]
-    context.Message(f'Checking {compiler} {flag}')
-    context.env.Replace(CCFLAGS=flags)
-    ret = context.TryCompile("""
-int main() {
-    return 0;
-}
-""", ext)
-    context.Result(ret)
-    return ret
+class BuildInfo():
+    """A utility class to save build information"""
 
+    def __init__(self):
+        self.info = {}
 
-def check_flag(context, flag):
-    """Check C specific compiler flags"""
-    return check_flag_helper(context, context.env.get("CC"), ".c", flag)
+    def update(self, var, value):
+        """save a variable in the build info"""
+        self.info[var] = value
 
+    def save(self, filename):
+        """Create a file to store path information for a build"""
+        with open(filename, "w") as build_info:
+            json.dump(self.info, build_info, skipkeys=True, indent=2)
 
-def check_flag_cc(context, flag):
-    """Check C++ specific compiler flags"""
-    return check_flag_helper(context, context.env.get("CXX"), ".cpp", flag)
+    def gen_script(self, script_name):
+        """Generate a shell script to set PATH, LD_LIBRARY_PATH, and PREFIX variables"""
+        with open(script_name, "w") as script:
+            script.write('# Automatically generated by'
+                         + f' {sys.argv[0]} at {datetime.datetime.today()}\n\n')
 
+            lib_paths = []
+            paths = []
+            components = []
 
-def check_flags(env, config, key, value):
-    """Check and append all supported flags"""
-    if GetOption('help') or GetOption('clean'):
-        return
-    checked = []
-    cxx = env.get('CXX')
-    for flag in value:
-        if flag in checked:
-            continue
-        insert = False
-        if key == "CCFLAGS":
-            if config.CheckFlag(flag) and (cxx is None or config.CheckFlagCC(flag)):
-                insert = True
-        elif key == "CFLAGS":
-            if config.CheckFlag(flag):
-                insert = True
-        elif cxx is not None and config.CheckFlagCC(flag):
-            insert = True
-        if insert:
-            env.AppendUnique(**{key: [flag]})
-        checked.append(flag)
-
-
-def append_if_supported(env, **kwargs):
-    """Check and append flags for construction variables"""
-    cenv = env.Clone()
-    config = Configure(cenv, custom_tests={'CheckFlag': check_flag,
-                                           'CheckFlagCC': check_flag_cc})
-    for key, value in kwargs.items():
-        if key not in ["CFLAGS", "CXXFLAGS", "CCFLAGS"]:
-            env.AppendUnique(**{key: value})
-            continue
-        check_flags(env, config, key, value)
-
-    config.Finish()
+            for var in sorted(self.info.keys()):
+                if not isinstance(self.info[var], str):
+                    continue
+                if "PREFIX" not in var:
+                    continue
+                if self.info[var] == "/usr":
+                    continue
+                script.write(f'SL_{var}={self.info[var]}\n')
+                components.append(var)
+                path = os.path.join(self.info[var], "bin")
+                lib = os.path.join(self.info[var], "lib")
+                lib64 = os.path.join(self.info[var], "lib64")
+                if os.path.exists(path) and path not in paths:
+                    paths.insert(0, path)
+                if os.path.exists(lib) and lib not in lib_paths:
+                    lib_paths.insert(0, lib)
+                if os.path.exists(lib64) and lib64 not in lib_paths:
+                    lib_paths.insert(0, lib64)
+            script.write(f'SL_LD_LIBRARY_PATH={os.pathsep.join(lib_paths)}\n')
+            script.write(f'SL_PATH={os.pathsep.join(paths)}\n')
+            component_list = ' '.join(components)
+            script.write(f'SL_COMPONENTS="{component_list}"\n')
+            script.write(f'SL_BUILD_DIR={self.info["BUILD_DIR"]}\n')
+            script.write(f'SL_SRC_DIR={os.getcwd()}\n')
 
 
 def ensure_dir_exists(dirname, dry_run):
@@ -612,7 +590,6 @@ class PreReqComponent():
 
         self.__dry_run = GetOption('no_exec')
         self._add_options()
-        self.__env.AddMethod(append_if_supported, "AppendIfSupported")
         self.__require_optional = GetOption('require_optional')
         self._has_icx = False
         self.download_deps = False
@@ -894,9 +871,10 @@ class PreReqComponent():
         # the number of go procs will be multiplied by jobs_opt.
         self.__env["ENV"]["GOMAXPROCS"] = "1"
 
-    def get_build_info(self):
-        """Retrieve the BuildInfo"""
-        return self.__build_info
+    def save_build_info(self):
+        """Save build info to file for later use"""
+        self.__build_info.gen_script('.build_vars.sh')
+        self.__build_info.save('.build_vars.json')
 
     def _add_options(self):
         """Add common options to environment"""
