+++ utils/docker/Dockerfile.leap.15
@@ -1,4 +1,4 @@
-# Copyright 2018-2022 Intel Corporation
+# Copyright 2018-2023 Intel Corporation
 # All rights reserved.
 #
 # 'recipe' for Docker to build an image of Leap based
@@ -67,9 +67,7 @@ RUN python3 -m pip --no-cache-dir install --upgrade pip && \
 
 WORKDIR /home/daos/pre
 COPY --chown=daos_server:daos_server SConstruct .
-COPY --chown=daos_server:daos_server site_scons/stack_analyzer.py site_scons/stack_analyzer.py
 COPY --chown=daos_server:daos_server site_scons/prereq_tools site_scons/prereq_tools
-COPY --chown=daos_server:daos_server site_scons/build_info site_scons/build_info
 COPY --chown=daos_server:daos_server site_scons/components site_scons/components
 COPY --chown=daos_server:daos_server utils/build.config utils/
