+++ utils/sl/fake_scons/SCons/Script/__init__.py
@@ -1,5 +1,5 @@
 """Fake scons environment shutting up pylint on SCons files"""
-# Copyright 2016-2022 Intel Corporation
+# Copyright 2016-2023 Intel Corporation
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -84,6 +84,10 @@ class DefaultEnvironment():
     def Zip(self, *_args, **_kw):
         """Fake Zip"""
         return []
+
+    def Tool(self, name):
+        """Fake Tool"""
+        return
 
     def CXXFile(self, *_args, **_kw):
         """Fake CXXFile"""
@@ -212,6 +216,7 @@ class DefaultEnvironment():
         return
 
     def __index__(self):
+        """Allow indexing"""
         return 0
 
     def Install(self, *_args, **_kw):
@@ -230,7 +235,7 @@ class DefaultEnvironment():
         return []
 
     def SConscript(self, s_dir):
-        "Fake SConscript"
+        """Fake SConscript"""
         return
 
     def Replace(self, *_args, **_kw):
@@ -299,12 +304,10 @@ class DefaultEnvironment():
 
     def d_program(self, *_args, **_kw):
         """Fake d_program"""
-
         return self.Program(*_args, **_kw)
 
     def d_test_program(self, *_args, **_kw):
         """Fake d_test_program"""
-
         return self.d_program(*_args, **_kw)
 
     def d_static_library(self, *_args, **_kw):
@@ -314,6 +317,10 @@ class DefaultEnvironment():
     def d_library(self, *_args, **_kw):
         """Fake d_library"""
         return self.Library(*_args, **_kw)
+
+    def analyze_setup(self, prefix, args):
+        """Fake analyze_setup"""
+        return
 
     def compiler_setup(self):
         """Fake compiler_setup"""
@@ -348,6 +355,7 @@ class Variables():
 
 class Configure():
     """Fake Configure"""
+
     def __init__(self, *_args, **_kw):
         self.env = DefaultEnvironment()
         """constructor"""
@@ -405,6 +413,7 @@ class Literal():
 
 class Dir():
     """Fake Dir"""
+
     def __init__(self, *_args, **_kw):
         self.abspath = os.getcwd()
         self.path = os.getcwd()
@@ -443,6 +452,7 @@ def SetOption(*_args, **_kw):
 
 class Help():
     """Fake Help"""
+
     def __init__(self, *_args, **_kw):
         """constructor"""
